core: mareco-high-bound-debug: implement computeInitialHighBound and re-organize MarecoAllowance in a more logical way
